Add tests as classpath resources to make project self-contained
Avoid downloading tests from third-party repositories and include them
as classpath resources inside the project/jar. Most if not all users
will want to run tests from within this set.

Add tests to verify that test files can be picked up and run with
the default JDBC executor.

Add dependency to HSQLDB to be able to execute the JDBC driver;
requires the jdk8 classifier cause this project compiles and runs
in JDK8. The default jar is compiled in JDK 11 so running it with JDK8
fails for various reasons.

Close hydromatic/sql-logic-test#11